--story=1005178 --user=lvzk Tagent注册 https://www.tapd.cn/54247054/s/1176368

+++ src/main/java/codedriver/module/autoexec/api/job/AutoexecJobConsoleLogTailApi.java
@@ -7,7 +7,7 @@ package codedriver.module.autoexec.api.job;
 
 import codedriver.framework.auth.core.AuthAction;
 import codedriver.framework.autoexec.auth.AUTOEXEC_BASE;
-import codedriver.framework.autoexec.dto.RunnerVo;
+import codedriver.framework.dto.runner.RunnerVo;
 import codedriver.framework.autoexec.dto.job.AutoexecJobVo;
 import codedriver.framework.autoexec.exception.AutoexecJobNotFoundException;
 import codedriver.framework.autoexec.exception.AutoexecJobRunnerNotFoundException;
@@ -19,7 +19,7 @@ import codedriver.framework.restful.annotation.Param;
 import codedriver.framework.restful.constvalue.OperationTypeEnum;
 import codedriver.framework.restful.core.privateapi.PrivateApiComponentBase;
 import codedriver.module.autoexec.dao.mapper.AutoexecJobMapper;
-import codedriver.module.autoexec.dao.mapper.RunnerMapper;
+import codedriver.framework.dao.mapper.runner.RunnerMapper;
 import codedriver.module.autoexec.service.AutoexecJobActionService;
 import codedriver.module.autoexec.service.AutoexecJobService;
 import com.alibaba.fastjson.JSONObject;
@@ -74,7 +74,7 @@ public class AutoexecJobConsoleLogTailApi extends PrivateApiComponentBase {
     })
     @Override
     public Object myDoService(JSONObject paramObj) throws Exception {
-        Integer runnerId = paramObj.getInteger("runnerId");
+        Long runnerId = paramObj.getLong("runnerId");
         Long jobId = paramObj.getLong("jobId");
         AutoexecJobVo jobVo = autoexecJobMapper.getJobInfo(jobId);
         if (jobVo == null) {
